ref: move edit table utils to parent columnCreation utils
